backend/pyproject.toml
[project]
name = "bill-parser-api"
version = "0.1.0"
description = "FastAPI backend for bill parsing"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "fastapi-limiter>=0.1.6",
    "google-genai>=1.11.0",
    "langchain>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-google-genai>=2.1.5",
    "pillow>=11.2.1",
    "python-decouple>=3.8",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "uvicorn>=0.34.2",
]
